Extract electron mirror check and path annotator
