+++ packages/shared-frontend-utils/src/formatUtil.ts
@@ -1,5 +1,4 @@
-import type { ResultItem } from '@/schemas/apiSchema'
-import type { operations } from '@/types/comfyRegistryTypes'
+import type { operations } from '@comfyorg/registry-types'
 
 export function formatCamelCase(str: string): string {
   // Check if the string is camel case
@@ -193,27 +192,6 @@ export function isValidUrl(url: string): boolean {
   } catch {
     return false
   }
-}
-const hasAnnotation = (filepath: string): boolean =>
-  /\[(input|output|temp)\]/i.test(filepath)
-
-const createAnnotation = (filepath: string, rootFolder = 'input'): string =>
-  !hasAnnotation(filepath) && rootFolder !== 'input' ? ` [${rootFolder}]` : ''
-
-const createPath = (filename: string, subfolder = ''): string =>
-  subfolder ? `${subfolder}/${filename}` : filename
-
-/** Creates annotated filepath in format used by folder_paths.py */
-export function createAnnotatedPath(
-  item: string | ResultItem,
-  options: { rootFolder?: string; subfolder?: string } = {}
-): string {
-  const { rootFolder = 'input', subfolder } = options
-  if (typeof item === 'string')
-    return `${createPath(item, subfolder)}${createAnnotation(item, rootFolder)}`
-  return `${createPath(item.filename ?? '', item.subfolder)}${
-    item.type ? createAnnotation(item.type, rootFolder) : ''
-  }`
 }
 
 /**
